# Enable the union feature of smallvec for better performance