--- a/Cargo.toml
+++ b/Cargo.toml
@@ -20,7 +20,7 @@ default = []
 reduced-trie = []
 
 [dev-dependencies]
-criterion = "0.3.5"
+criterion = "0.4.0"
 rand = "0.8.4"
 
 [[bench]]
@@ -33,4 +33,4 @@ required-features = []
 all-features = true
 
 [dependencies]
-smallvec = "1.6.1"
+smallvec = { version = "1.6.1", features = ["union"] }