`union` feature was stabilized since Rust 1.49.0.

```
cedar build             time:   [1.3783 µs 1.3836 µs 1.3897 µs]
                        change: [-1.0152% -0.5505% -0.0928%] (p = 0.03 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

cedar exact_match_search
                        time:   [1.4028 µs 1.4044 µs 1.4061 µs]
                        change: [-1.2937% -1.0430% -0.8104%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 9 outliers among 100 measurements (9.00%)
  5 (5.00%) low mild
  1 (1.00%) high mild
  3 (3.00%) high severe

cedar common_prefix_search
                        time:   [1.4298 µs 1.4316 µs 1.4336 µs]
                        change: [-1.7556% -1.3444% -0.9916%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 10 outliers among 100 measurements (10.00%)
  9 (9.00%) high mild
  1 (1.00%) high severe

cedar common_prefix_predict
                        time:   [1.4905 µs 1.4930 µs 1.4957 µs]
                        change: [-3.0412% -2.6523% -2.2536%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  7 (7.00%) high mild
  3 (3.00%) high severe
```